+++ home/tests/test_assessment_import_export.py
@@ -370,10 +370,15 @@ class TestImportExportRoundtrip:
                 {
                     "type": "categorical_question",
                     "value": {
-                        "semantic_id": "unique-id",
                         "question": "test question",
                         "error": "test error",
-                        "answers": [{"answer": "test answer", "score": 2.0}],
+                        "answers": [
+                            {
+                                "answer": "test answer",
+                                "score": 2.0,
+                                "semantic_id": "unique_id",
+                            }
+                        ],
                     },
                 }
             ],